FIX: issue #64 formatargspec deprecation (2)
