+++ mdp/signal_node.py
@@ -19,20 +19,7 @@ import inspect
 
 import mdp
 from mdp import numx
-
-def _legacysignature(signature):
-    """
-    For retrocompatibility reasons, we don't use a standard Signature.
-    Instead, we use the string generated by this method.
-    Basically, from a Signature we create a string and remove the default values.
-    """
-    listsignature = str(signature)[1:-1].split(",")
-    for counter, param in enumerate(listsignature):
-        if param.count("=") > 0:
-            listsignature[counter] = param[0:param.index("=")].strip()
-        else:
-            listsignature[counter] = param.strip()
-    return ", ".join(listsignature)
+from mdp.utils import _legacysignature
 
 class NodeException(mdp.MDPException):
     """Base class for exceptions in `Node` subclasses."""
@@ -187,7 +174,14 @@ class NodeMetaclass(type):
             argnames.append(varargs)
         if varkwargs:
             argnames.append(varkwargs)
-        signature = _legacysignature(inspect.signature(func))
+        try:
+            signature = _legacysignature(inspect.signature(func))
+        except:
+            signature = inspect.formatargspec(regargs,	      
+                                          varargs,	
+                                          varkwargs,	
+                                          defaults,	
+                                          formatvalue=lambda value: "")[1:-1]
         return dict(name=func.__name__,
                     signature=signature,
                     argnames=argnames,
