+++ mdp/utils/routines.py
@@ -11,6 +11,22 @@ numx_description = mdp.numx_description
 import random
 import itertools
 
+def _legacysignature(signature):
+    """
+    Used to replace inspect.formatargspec by inspect.signature.
+    
+    For retrocompatibility reasons, we don't use a standard Signature.
+    Instead, we use the string generated by this method.
+    Basically, from a Signature we create a string and remove the default values.
+    """
+    listsignature = str(signature)[1:-1].split(",")
+    for counter, param in enumerate(listsignature):
+        if param.count("=") > 0:
+            listsignature[counter] = param[0:param.index("=")].strip()
+        else:
+            listsignature[counter] = param.strip()
+    return ", ".join(listsignature)
+
 def timediff(data):
     """Returns the array of the time differences of data."""
     # this is the fastest way we found so far
